+++ angular-ui/src/app/app.routes.ts
@@ -6,6 +6,10 @@ export const routes: Routes = [
     loadComponent: () => import('./feature/home/home.component').then(m => m.HomeComponent),
   },
   {
+    path: 'products',
+    loadComponent: () => import('./feature/product/product-list/product-list.component').then(m => m.ProductListComponent),
+  },
+  {
     path: 'auth',
     loadChildren: () => import('./feature/auth/auth.routes').then(m => m.AUTH_ROUTES),
   },